fix: integrate execute graphql with the dapp
