+++ libs/tangle-shared-ui/src/utils/executeGraphQL.ts
@@ -1,10 +1,12 @@
-import type { TypedDocumentString } from '../graphql/graphql';
+import type { NetworkType, TypedDocumentString } from '../graphql/graphql';
 
 export async function executeGraphQL<TResult, TVariables>(
-  endpoint: string,
+  network: NetworkType,
   query: TypedDocumentString<TResult, TVariables>,
   ...[variables]: TVariables extends Record<string, never> ? [] : [TVariables]
 ) {
+  const endpoint = getGraphQLEndpointFromNetwork(network);
+
   const response = await fetch(endpoint, {
     method: 'POST',
     headers: {
@@ -23,3 +25,14 @@ export async function executeGraphQL<TResult, TVariables>(
 
   return response.json() as Promise<{ data: TResult }>;
 }
+
+const getGraphQLEndpointFromNetwork = (network: NetworkType) => {
+  switch (network) {
+    case 'MAINNET':
+      return 'https://mainnet-gql.tangle.tools/graphql';
+    case 'TESTNET':
+      return 'https://testnet-gql.tangle.tools/graphql';
+    default:
+      throw new Error(`Invalid network: ${network}`);
+  }
+};
